+++ utils/app/settings.ts
@@ -1,19 +1,19 @@
-import {Settings} from "@/types/settings"
+import {Theme} from "@/types/settings"
 
-export const STORAGE_KEY_SETTINGS = "settings"
+export const STORAGE_KEY_THEME = "theme"
 export const STORAGE_KEY_API_KEY = "apiKey"
 export const STORAGE_KEY_UNLOCK_CODE = "unlockCode"
 export const STORAGE_KEY_SHOW_CHAT_BAR = "showChatBar"
 export const STORAGE_KEY_SHOW_PROMPT_BAR = "showPromptBar"
 
-export const getSettings = (): Settings => {
-  let settings: Settings = {
+export const getSettings = (): Theme => {
+  let settings: Theme = {
     theme: "dark"
   }
-  const settingsAsString = localStorage.getItem(STORAGE_KEY_SETTINGS)
+  const settingsAsString = localStorage.getItem(STORAGE_KEY_THEME)
   if (settingsAsString) {
     try {
-      let savedSettings = JSON.parse(settingsAsString) as Settings
+      let savedSettings = JSON.parse(settingsAsString) as Theme
       settings = Object.assign(settings, savedSettings)
     } catch (error) {
       console.error(`Local storage error: ${error}`)
@@ -21,12 +21,12 @@ export const getSettings = (): Settings => {
   }
   return settings
 }
-
-export const saveSettings = (settings: Settings) => {
-  localStorage.setItem(STORAGE_KEY_SETTINGS, JSON.stringify(settings))
+export const saveTheme = (theme: Theme) => {
+  localStorage.setItem(STORAGE_KEY_THEME, JSON.stringify(theme))
 }
+export const removeTheme = () => localStorage.removeItem(STORAGE_KEY_THEME)
 
-export const removeSettings = () => localStorage.removeItem(STORAGE_KEY_SETTINGS)
+export const removeSettings = () => localStorage.removeItem(STORAGE_KEY_THEME)
 export const getApiKey = () => localStorage.getItem(STORAGE_KEY_API_KEY)
 export const saveApiKey = (apiKey: string) => localStorage.setItem(STORAGE_KEY_API_KEY, apiKey)
 export const removeApiKey = () => localStorage.removeItem(STORAGE_KEY_API_KEY)